+++ lib/components/PrettyIdentifier.vue
@@ -1,5 +1,5 @@
 <script setup>
-import useClipboard from 'vue-clipboard3'
+import CopyableText from './CopyableText.vue'
 </script>
 
 <script>
@@ -10,23 +10,9 @@ export default {
             required: true,
         },
     },
-
-    methods: {
-
-        async copyIdentifier() {
-            const { toClipboard } = useClipboard()
-            await toClipboard(this.value)
-        },
-    },
 }
 </script>
 
 <template>
-  <span>
-    {{ value }}
-    <a v-if="value"
-       href="#" @click.prevent="copyIdentifier()">
-      <o-icon icon="copy" />
-    </a>
-  </span>
+  <CopyableText :text="value" />
 </template>